+++ tests/test_scope_file.py
@@ -17,8 +17,8 @@ def file_idfn(path):
 
 
 # generate filenames to import
-files = list(Path(".").rglob("**/data/tc3_scope_*.csv"))
-# files = list(Path(".").rglob("**/data/vs_eng*"))
+files = sorted(Path(".").rglob("**/data/tc3_scope_*.csv.gz*"))
+files_pyarrow = [file for file in files if "noOS" in file.name]
 
 # list of files broken for loading:
 broken_load = ["Seperator_1", "Seperator_4"]
@@ -29,6 +29,21 @@ def filenames(request):
     return request.param
 
 
+@pytest.fixture(scope="module", params=files_pyarrow, ids=file_idfn)
+def filenames_pyarrow(request):
+    return request.param
+
+
+def _get_as_buffer(file):
+    """Convert file into buffer."""
+    if file.suffix in [".gz", ".gzip"]:
+        with open(file, "rb") as f:
+            return BytesIO(f.read())
+    else:
+        with open(file, "rt") as f:
+            return StringIO(f.read())
+
+
 # Testing files in data
 class TestScopeFile:
     @staticmethod
@@ -36,27 +51,47 @@ class TestScopeFile:
     @pytest.mark.parametrize("backend", ["pandas"])
     @pytest.mark.parametrize("use_buffer", [False, True])
     def test_scope_file(filenames, backend, native_dtypes, use_buffer):
-        if native_dtypes & (backend == "datatable"):
-            pytest.skip("unsupported configuration")
+        file = filenames
 
-        if any(sep in str(filenames) for sep in broken_load):
-            with pytest.raises(ValueError):
-                ScopeFile(filenames)
-            return None
+        if native_dtypes & (backend == "datatable"):
+            pytest.skip("unsupported configuration (datatable + native)")
+
+        if any(sep in str(file) for sep in broken_load):
+            pytest.skip("unsupported file format (separators)")
 
         if use_buffer:
-            if filenames.suffix in [".gz", ".gzip"]:
-                with open(filenames, "rb") as f:
-                    filenames = BytesIO(f.read())
-            else:
-                with open(filenames, "rt") as f:
-                    filenames = StringIO(f.read())
+            file = _get_as_buffer(file)
 
-        sf = ScopeFile(filenames)
+        sf = ScopeFile(file)
         for c in sf:
             assert sf[c].info
 
         sf.load(native_dtypes=native_dtypes, backend=backend)
+
+        assert len(sf._get_data_cols()) == len(sf._channels)
+        assert len(np.unique(sf._get_time_cols())) + len(sf._get_data_cols()) == len(
+            sf._data
+        )
+        assert all([type(v) == np.ndarray for v in sf._data.values()])
+
+        # monotonic time
+        for c in sf:
+            assert np.allclose(np.diff(sf[c].time), sf[c].sample_time)
+
+    @staticmethod
+    @pytest.mark.parametrize("native_dtypes", [False, True])
+    @pytest.mark.parametrize("use_buffer", [False, True])
+    def test_pyarrow_backend(filenames_pyarrow, native_dtypes, use_buffer):
+        file = filenames_pyarrow
+
+        if use_buffer:
+            file = _get_as_buffer(file)
+
+        sf = ScopeFile(file)
+        for c in sf:
+            assert sf[c].info
+
+        sf.load(native_dtypes=native_dtypes, backend="pyarrow")
 
         assert len(sf._get_data_cols()) == len(sf._channels)
         assert len(np.unique(sf._get_time_cols())) + len(sf._get_data_cols()) == len(